Add Task 2 - v1

+++ 0x13-bit_manipulation/1-print_binary.c
@@ -9,8 +9,8 @@
  */
 void print_binary(unsigned long int n)
 {
-	unsigned long int mask;
-	unsigned long int i;
+	int mask;
+	int i;
 
 	if (n == 0)
 		_putchar('0');
